requirements.txt
streamlit
pandas
numpy
pyotp
plotly
smartapi-python
